refactor: animation
