+++ next-docs/pages/index.tsx
@@ -1,18 +1,12 @@
 import React from 'react';
 
-import Image from 'next/image';
-
-import { useViewportScroll, useTransform } from 'framer-motion';
-
-import imageMoon from '../public/moon.png';
-import imageEarth from '../public/earth.png';
-
 import Future from '../components/facing/Future';
 import ForDesigners from '../components/facing/ForDesigners';
 import ForDevelopers from '../components/facing/ForDevelopers';
 import Contributors from '../components/facing/Contributors';
 import News from '../components/facing/News';
 import Logo from '../components/sidebar/Logo';
+import MoonAndEarthAnimation from '../components/facing/MoonAndEarth';
 
 const OpenSourceBadge = () => (
   <span className="block w-min whitespace-nowrap mb-8 text-active bg-active-80 uppercase tracking-widest py-2 px-4 rounded-lg font-bold">
@@ -59,31 +53,6 @@ const Footer = () => (
 );
 
 export default function Home() {
-  const { scrollYProgress } = useViewportScroll();
-  const earthScale = useTransform(
-    scrollYProgress,
-    (scrollYProgress) => 1 + scrollYProgress * 2
-  );
-  const moonScale = useTransform(
-    scrollYProgress,
-    (scrollYProgress) => 1 - scrollYProgress * 0.5
-  );
-
-  const [hookedEarthScale, setHookedEarthScale] = React.useState(1);
-  const [hookedMoonScale, setHookedMoonScale] = React.useState(1);
-  React.useEffect(() => {
-    const unsubscribeEarthScale = earthScale.onChange((v) =>
-      setHookedEarthScale(v)
-    );
-    const unsubscribeMoonScale = moonScale.onChange((v) =>
-      setHookedMoonScale(v)
-    );
-    return () => {
-      unsubscribeEarthScale();
-      unsubscribeMoonScale();
-    };
-  }, [earthScale, moonScale]);
-
   return (
     <>
       <OpenSourceBadge />
@@ -102,21 +71,7 @@ export default function Home() {
       <ForDesigners />
       <ForDevelopers />
 
-       {/* Moon background image */}
-      <div
-        style={{ transform: `scale(${hookedEarthScale})` }}
-        className="fixed right-0 top-300 z-20 w-5/12"
-      >
-        <Image src={imageEarth} alt="Earth" />
-      </div>
-
-      {/* Moon background image */}
-      <div
-        style={{ transform: `scale(${hookedMoonScale}` }}
-        className="fixed right-0 top-250 z-10 w-5/12"
-      >
-        <Image src={imageMoon} alt="Moon" />
-      </div>
+      <MoonAndEarthAnimation />
 
       <section className="mt-20">
         <Future />
